исправлено всякое.

+++ src/main/java/com/actiongroup/actionserver/services/users/UserService.java
@@ -5,6 +5,7 @@ import com.actiongroup.actionserver.models.users.Role;
 import com.actiongroup.actionserver.models.users.User;
 import com.actiongroup.actionserver.models.users.UserSettings;
 import com.actiongroup.actionserver.repositories.user.RoleRepository;
+import com.actiongroup.actionserver.repositories.user.UserRelationRepository;
 import com.actiongroup.actionserver.repositories.user.UserRepository;
 import com.actiongroup.actionserver.repositories.user.UserSettingsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,22 +28,27 @@ public class UserService implements UserDetailsService {
 
     @Autowired
     UserSettingsRepository settingsRepo;
+    @Autowired
+    UserRelationRepository relationRepository;
 
     public boolean save(User user) {
-        User userFromDB = userRepo.findByUsername(user.getUsername());
+        if(userRepo.existsByUsername(user.getUsername())) return false;
 
-
-
+        //User userFromDB = userRepo.findByUsername(user.getUsername());
         if(user.getRoles() == null || user.getRoles().size()==0){
             Role roles = roleRepo.findByName("ROLE_USER").get();
             user.setRoles(Collections.singleton(roles));
         }
-        UserSettings settings = new UserSettings();
-        settings.setVerified(false);
-        settings.setUser(user);
 
 
         userRepo.save(user);
+
+        UserSettings settings = settingsRepo.findByUser(user);
+        if(settings == null)
+            settings = new UserSettings();
+        settings.setVerified(false);
+        settings.setUser(user);
+
         settingsRepo.save(settings);
         return true;
     }
@@ -57,7 +63,10 @@ public class UserService implements UserDetailsService {
 
     public void deleteUser(User user){
         deleteUserSettings(settingsRepo.findByUser(user));
-        userRepo.deleteById(user.getId());
+        relationRepository.deleteAll(relationRepository.findBySourceUserOrTargetUser(user,user));
+        if(user!=null)
+            userRepo.deleteById(user.getId());
+
     }
 
     private void deleteUserSettings(UserSettings settings){
@@ -79,7 +88,6 @@ public class UserService implements UserDetailsService {
     public UserSettings getSettingsByUser(User user){
         return settingsRepo.findByUser(user);
     }
-
 
     @Override
     public org.springframework.security.core.userdetails.UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
